Update inference speed

+++ src/fine_tune.py
@@ -62,7 +62,7 @@ def load_model_and_tokenizer(model_dir, device):
     try:
         model = AutoModelForMaskedLM.from_pretrained(model_dir, trust_remote_code=True, dtype=torch.bfloat16)
     except:
-        model = AutoModelForMaskedLM.from_pretrained(model_dir, trust_remote_code=True, dtype=torch.bfloat16)
+        model = AutoModelForMaskedLM.from_pretrained(model_dir, trust_remote_code=True)
     tokenizer = AutoTokenizer.from_pretrained(model_dir, trust_remote_code=True)
     model.to(device)
     return model, tokenizer
